patternProject/patternProject/analysis/serializers.py
from .models import Analysis, Interaction
from rest_framework import serializers

class AnalysisSerializer(serializers.ModelSerializer):
    class Meta:
        model = Analysis
        fields = '__all__'


class InteractionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Interaction
        fields = '__all__'