patternProject/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Django==4.0.3
django-allauth==0.50.0
django-cors-headers==3.12.0
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
drf-jwt==1.19.2
idna==3.3
importlib-metadata==4.11.3
Markdown==3.3.6
mysqlclient==2.1.0
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
pycparser==2.21
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.9
zipp==3.7.0
